- Optimize logo asset (resize) to fix initial loading delay. - Minor visual adjustments for consistency.
