+++ lib/main.dart
@@ -1,4 +1,4 @@
-import 'package:example/view/attendence_view.dart';
+import 'package:example/view/attendance_view.dart';
 import 'package:example/view/home_view.dart';
 import 'package:example/view/login_page.dart';
 import 'package:flutter/material.dart';
