+++ lib/view/login_page.dart
@@ -1,4 +1,4 @@
-import 'package:example/view/attendence_view.dart';
+import 'package:example/view/attendance_view.dart';
 import 'package:example/view/home_view.dart';
 import 'package:example/view/widgets/custom_text_field.dart';
 import 'package:flutter/material.dart';
@@ -109,7 +109,7 @@ class _LogInPageState extends State<LogInPage> {
               padding: const EdgeInsets.all(8.0),
               child: GestureDetector(
                 onTap: () {
-                  Get.offAll(AttendenceView());
+                  Get.offAll(AttendanceView());
                 },
                 child: Container(
                   width: 120,
